Added functionality for filtering and emergency override
Added the functionality to filter tasks by priority and if they are completed. The tasks can be fetched by ID, can be updated and deleted. Very basic user assignment so that it can be tested and changed later. Emergency overrride is in place where the task takes precedent over non-emergency or completed emergency tasks.