Fixed Emergency Meetings displaying
Emergency meetings no longer tries to get the users of a project to display as attendees. The emergency meeting overlay now displays on the dashboard.
It also filters the overlay only if the user is assigned as an attendee will it display, it also allows only the user who created the meeting to be able to edit and cancel it.
Admins can see all meetings, edit and cancel them.